Intent: modularize dynamics facade Operators involved: [Coupling, Self-organization] Affected invariants: [#1, #7] Key changes: - introduce coordination, adaptation, selectors, runtime submodules and central aliases facade for dynamics - reduce tnfr.dynamics.__init__ to API re-export layer and adjust typing stubs - update runtime orchestration to consult facade for patchable hooks while preserving phase/vf coordination helpers Expected risks/dissonances: refactor touches widely used module, so import regressions may emerge despite targeted tests Metrics: C(t) semantics preserved; no change to νf, phase traces expected Equivalence map: tnfr.dynamics.* functions now delegate to submodules (see new module layout)
